頂点インデックスのフォーマット少し修正

+++ GgafDx/include/jp/ggaf/dx/util/XFile/framework/Frm_3DMath.h
@@ -136,7 +136,7 @@ struct Array {
 };
 
 //typedef Array<unsigned short, 3> Face;
-typedef Array<unsigned short, 4> Face; //modify tsuge
+typedef Array<uint32_t, 4> Face; //modify tsuge
 
 typedef Array<float, 3> Vertex;
 typedef Array<float, 2> TCoord;
